+++ test/PHPUnit/connecteur/s2low/S2lowActesTest.php
@@ -3,84 +3,74 @@
 class S2lowActesTest extends PastellTestCase
 {
     /**
-     * @throws Exception
+     * @throws DonneesFormulaireException
+     * @throws S2lowException
      */
-    public function testPostActesOK()
+    public function testPostActesOK(): void
     {
-        $info = $this->createConnector('s2low', "S2LOW", self::ID_E_COL);
-        $id_ce = $info['id_ce'];
-        $this->associateFluxWithConnector($id_ce, "actes-generique", "TdT", self::ID_E_COL);
+        $curlWrapper = $this->createMock(CurlWrapper::class);
 
-        $connecteurDonnerFormulaire = $this->getDonneesFormulaireFactory()->getConnecteurEntiteFormulaire($id_ce);
-
-        $connecteurDonnerFormulaire->addFileFromCopy(
-            "classification_file",
-            "classification.xml",
-            __DIR__ . "/fixtures/classification-exemple.xml"
-        );
-
-
-        $info = $this->createDocument('actes-generique', self::ID_E_COL);
-
-        $id_d = $info['id_d'];
-
-        $donneesFormulaire = $this->getDonneesFormulaireFactory()->get($id_d);
-        $donneesFormulaire->addFileFromData("arrete", "test.pdf", "aaaa");
-        $donneesFormulaire->addFileFromData("autre_document_attache", "annexe1.pdf", "foo");
-
-        $donneesFormulaire->setTabData(
-            [
-                'acte_nature' => '3',
-                'numero_de_lacte' => '201903251130',
-                'objet' => 'TEST',
-                'date_de_lacte' => '2019-03-25',
-                'classification' => '2.1'
-            ]
-        );
-
-        $curlWrapper = $this->getMockBuilder('CurlWrapper')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $curlWrapper->expects($this->any())
+        $curlWrapper
             ->method('get')
-            ->will($this->returnCallback(function ($url) {
-                if ($url == "/modules/actes/actes_classification_fetch.php?api=1") {
-                    return file_get_contents(__DIR__ . "/fixtures/classification-exemple.xml");
-                } elseif ($url == "/admin/users/api-list-login.php") {
+            ->willReturnCallback(function ($url) {
+                if ($url === '/modules/actes/actes_classification_fetch.php?api=1') {
+                    return file_get_contents(__DIR__ . '/fixtures/classification-exemple.xml');
+                }
+                if ($url === '/admin/users/api-list-login.php') {
                     return true;
-                } elseif ($url == "/modules/actes/actes_transac_create.php") {
+                }
+                if ($url === '/modules/actes/actes_transac_create.php') {
                     return "OK\n666";
                 }
-                throw new Exception("$url inatendu");
-            }));
+                throw new \RuntimeException("$url inatendu");
+            });
 
         $addPostDataCall = [];
 
-        $curlWrapper->expects($this->any())
+        $curlWrapper
             ->method('addPostData')
-            ->will($this->returnCallback((function ($key, $value) use (&$addPostDataCall) {
-                $addPostDataCall[$key] = $value;
-                return true;
-            })));
+            ->willReturnCallback(
+                (static function ($key, $value) use (&$addPostDataCall) {
+                    $addPostDataCall[$key] = $value;
+                    return true;
+                })
+            );
 
-        $curlWrapperFactory = $this->getMockBuilder('CurlWrapperFactory')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $curlWrapperFactory = $this->createMock(CurlWrapperFactory::class);
 
-        $curlWrapperFactory->expects($this->any())
+        $curlWrapperFactory
             ->method('getInstance')
             ->willReturn($curlWrapper);
 
+        $this->getObjectInstancier()->setInstance(CurlWrapperFactory::class, $curlWrapperFactory);
 
-        $this->getObjectInstancier()->setInstance('CurlWrapperFactory', $curlWrapperFactory);
+        $form = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
+        $form->addFileFromCopy(
+            'classification_file',
+            'classification.xml',
+            __DIR__ . '/fixtures/classification-exemple.xml'
+        );
 
-        /** @var S2low $s2low */
-        $s2low = $this->getConnecteurFactory()->getConnecteurById($id_ce);
+        $s2low = new S2low($this->getObjectInstancier());
+        $s2low->setConnecteurConfig($form);
 
-        $this->assertTrue($s2low->postActes($donneesFormulaire));
+        $acte = new TdtActes();
+        $acte->acte_nature = '3';
+        $acte->numero_de_lacte = '201903251130';
+        $acte->objet = 'TEST';
+        $acte->date_de_lacte = '2019-03-25';
+        $acte->classification = '2.1';
+        $acte->arrete = new Fichier();
+        $acte->arrete->filepath = __DIR__ . '/fixtures/classification-exemple.xml';
+        $acte->arrete->filename = 'test.pdf';
 
-        $this->assertEquals("666", $donneesFormulaire->get('tedetis_transaction_id'));
+        $annexe = new Fichier();
+        $annexe->filepath = __DIR__ . '/fixtures/classification-exemple.xml';
+        $annexe->filename = 'annexe1.pdf';
+
+        $acte->autre_document_attache = [$annexe];
+
+        $this->assertSame('666', $s2low->sendActes($acte));
 
         $this->assertEquals(
             [
@@ -91,11 +81,9 @@ class S2lowActesTest extends PastellTestCase
                 'decision_date' => '2019-03-25',
                 'en_attente' => 0,
                 'document_papier' => 0,
-                'type_acte' => '99_AI',
-                'type_pj[]' => '99_AI',
                 'classif1' => '2',
                 'classif2' => '1',
-             ],
+            ],
             $addPostDataCall
         );
     }